+++ src/epblas-parallel/kind16/xtrmv.c
@@ -48,13 +48,13 @@ static inline TC cconj(TC z) { return conjq(z); }
  * qtrmv_blkrange (port of the openblas trmv_thread sqrt-partition). Work per
  * column j is proportional to j, so the narrow slab lands at the heavy high
  * columns. range[0..nthreads] are column boundaries over [0,n). */
-static void xtrmv_blkrange(ptrdiff_t n, int nthreads, ptrdiff_t *range)
+static void xtrmv_blkrange(ptrdiff_t n, ptrdiff_t nthreads, ptrdiff_t *range)
 {
     const ptrdiff_t mask = 7;
     const double dnum = (double)n * (double)n / (double)nthreads;
     range[nthreads] = n;
     ptrdiff_t i = 0;
-    int nc = 0;
+    ptrdiff_t nc = 0;
     while (i < n && nc < nthreads) {
         ptrdiff_t width;
         if (nthreads - nc > 1) {
@@ -72,7 +72,7 @@ static void xtrmv_blkrange(ptrdiff_t n, int nthreads, ptrdiff_t *range)
         nc++;
         i += width;
     }
-    for (int t = 0; t < nthreads - nc; ++t) range[t] = range[nthreads - nc];
+    for (ptrdiff_t t = 0; t < nthreads - nc; ++t) range[t] = range[nthreads - nc];
 }
 
 /* Threaded out-of-place path (incx==1). Returns 1 if handled, 0 to fall back to
@@ -96,7 +96,7 @@ static bool xtrmv_omp(bool upper, bool trans_t, bool conj_a, bool nounit, ptrdif
         ptrdiff_t *rng = NULL;
         if (upper && n >= XTRMV_BLOCK_MIN) {
             rng = (ptrdiff_t *)malloc(((size_t)nthreads + 1) * sizeof(ptrdiff_t));
-            if (rng) xtrmv_blkrange(n, (int)nthreads, rng);
+            if (rng) xtrmv_blkrange(n, nthreads, rng);
         }
         const bool blk = (rng != NULL);
         #pragma omp parallel num_threads(nthreads)
